fix: get web client building, syncing messages, and loading contacts
Bundles the fixes from an extended flutter run -d web-server debugging
pass — the web client wasn't compiling cleanly, wasn't loading message
history, and wasn't showing contact names/avatars.

Build/compile:
- Replace `dart:io` imports with `package:universal_io/io.dart` in
  database.dart, startup_tasks.dart, incoming_message_handler.dart,
  socket_service.dart, user_certificates.dart, and lifecycle_service.dart
  so these files compile for web instead of failing on an unavailable
  dart:io API.

ObjectBox web stubs:
- Give `ToOne`/`ToMany` (lib/database/html/objectbox.dart) real in-memory
  relation storage instead of always-null/no-op stubs. Several models
  silently broke without it: Attachment.message, Message.chat,
  Message.handleRelation, Handle.contactsV2, ContactV2.handles.
- Flesh out the web Chat/Handle/Message/Attachment models to match their
  io/ counterparts (saveAsync/toggle* methods, sort(), isNewerThan(),
  displayWidth/Height/displayBox, contact-aware displayName/initials,
  wallpaper fields, etc.) so shared action/UI code that calls these no
  longer throws on web.

Message loading (root cause of the original "messages never load" bug):
- SyncInterface.bulkSyncData() now parses messages directly from the raw
  map on web instead of hydrating them from the local ObjectBox boxes,
  which are never initialized there (LateInitializationError on every
  call). This was causing _loadMoreMessages to fail and retry forever.

Contacts:
- Route contact sync through the server-fetch path on web (alongside
  desktop) in ContactV2Actions, since flutter_contacts has no web support
  and would otherwise crash instead of populating anything.

Setup/sync completion:
- SyncProgress's web branch now actually finishes setup (sets
  finishedSetup, runs ChatsSvc.init, StartupTasks.onStartup,
  NetworkTasks.onConnect) — the native full-sync completion flow it
  depended on never runs on web, so finishedSetup stayed false forever
  and ChatsSvc kept bailing out early.
- Await the real server-details fetch before the web minimum-version
  compatibility check in main.dart, instead of reading a cache that's
  still empty at that point.

Query layer:
- Split chat_stats_queries.dart and search_query_helper.dart into
  io/web conditional-import pairs — both were built entirely on
  ObjectBox query APIs unavailable on web.

Isolates:
- Run isolate requests (GlobalIsolate.send/broadcast) directly on the
  main thread on web, since dart:isolate doesn't exist there.

Misc fixes found along the way:
- Guard ChatsService.init() against concurrent calls racing on shared
  state.
- Guard CustomGroupsService against running on web (no local DB there).
- Null-safety fix in a ChatActions log line, `as int` cast fixes in two
  migration files, and an unused import removed from storage_actions.dart.

Debugging tooling:
- Add a dev-only JS error forwarder to web/index.html, gated behind
  kDebugMode, that ships console errors/warnings to a local collector —
  makes iterating on web bugs practical without manually copy-pasting
  DevTools output on every reload.

+++ lib/database/migrations/chat_latest_message_migration.dart
@@ -60,7 +60,7 @@ class ChatLatestMessageMigration {
         chatBox.putMany(chatsToUpdate);
       }
 
-      processed += batch.length;
+      processed += batch.length as int;
       // the await is what gets this to the splash while the migration owns the isolate.
       await StartupTasks.setSplashProgress(processed / totalChats);
       Logger.info("Backfilled $processed / $totalChats chats...", tag: "DB-Migration");
